# swagger.yaml
openapi: 3.0.3
info:
    title: Coffee Supply Management System API
    version: "1.0.0"
    contact:
        name: Team 67 - PZSP2
        email: coffee-system@pw.edu.pl
...
paths:
...
    /distributors:
        get:
            tags: [Distributors]
            summary: List distributors
            operationId: listDistributors
            security:
                - BearerAuth: []
            responses:
                "200":
                    description: List of distributors
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/DistributorResponse"
                "401":
                    $ref: "#/components/responses/Unauthorized"

        post:
            tags: [Distributors]
            summary: Add distributor
            operationId: createDistributor
            security:
                - BearerAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/DistributorCreateRequest"
                        example:
                            username: "Kaffa Premium Sp. z o.o."
                            contact_email: "orders@kaffa.pl"
                            contact_phone: "+48 123 456 789"
                            daily_prices:
                                - day: 1
                                  base_price: 12.0
                                  availability_kg: 200.0
                                  discount_tiers:
                                      - level: 1
                                        quantity_kg: 30.0
                                        unit_price: 10.0
                                      - level: 2
                                        quantity_kg: 60.0
                                        unit_price: 8.0
                            delivery_params:
                                - building_id: "bld-001"
                                  lead_time_days: 1
                                  fixed_cost_pln: 50.0
                                  correction_cost_per_kg: 1.0
                                  max_correction_kg: 20.0
            responses:
                "201":
                    description: Distributor created
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DistributorResponse"
    /distributors/{distributor_id}:
        get:
            tags: [Distributors]
            summary: Distributor details
            operationId: getDistributor
            security:
                - BearerAuth: []
            parameters:
                - $ref: "#/components/parameters/DistributorId"
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DistributorResponse"
                "404":
                    $ref: "#/components/responses/NotFound"

        put:
            tags: [Distributors]
            summary: Update distributor
            operationId: updateDistributor
            security:
                - BearerAuth: []
            parameters:
                - $ref: "#/components/parameters/DistributorId"
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/DistributorUpdateRequest"
            responses:
                "200":
                    description: Updated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DistributorResponse"
                "404":
                    $ref: "#/components/responses/NotFound"

        delete:
            tags: [Distributors]
            summary: Delete distributor
            operationId: deleteDistributor
            security:
                - BearerAuth: []
            parameters:
                - $ref: "#/components/parameters/DistributorId"
            responses:
                "204":
                    description: Deleted
                "404":
                    $ref: "#/components/responses/NotFound"
    /distributors/self/prices:
        put:
            tags: [Distributors - Self Service]
            summary: Update own price list (distributor via API key)
            operationId: updateOwnPrices
            security:
                - ApiKeyAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/DistributorUpdateRequest"
            responses:
                "200":
                    description: Price list updated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/DistributorResponse"
                "401":
                    $ref: "#/components/responses/Unauthorized"
...
    /inventory:
        get:
            tags: [Inventory]
            summary: Current inventory levels
            operationId: getInventory
            security:
                - BearerAuth: []
            responses:
                "200":
                    description: Inventory levels for all buildings
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/InventoryStatus"
    /inventory/{building_id}:
        put:
            tags: [Inventory]
            summary: Update building inventory
            operationId: updateInventory
            security:
                - BearerAuth: []
            parameters:
                - $ref: "#/components/parameters/BuildingId"
                - name: current_kg
                  in: query
                  required: true
                  schema:
                      type: number
                      format: float
                      minimum: 0
            responses:
                "200":
                    description: Inventory updated
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    building_id:
                                        type: string
                                    current_inventory_kg:
                                        type: number
...
components:
    securitySchemes:
        BearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: Coordinator session token (from `/auth/login`)
        ApiKeyAuth:
            type: apiKey
            in: header
            name: X-Api-Key
            description: Distributor API key (format `cof_...`)
    parameters:
        DistributorId:
            name: distributor_id
            in: path
            required: true
            schema:
                type: string
                format: UUID
            description: Distributor UUID
        BuildingId:
            name: building_id
            in: path
            required: true
            schema:
                type: string
                format: UUID
            description: Building UUID
    responses:
        Unauthorized:
            description: Missing or invalid token/API key
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
                    example:
                        detail: "Authentication required"
        NotFound:
            description: Resource not found
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/ErrorResponse"
                    example:
                        detail: "Not found"
    schemas:
        ErrorResponse:
            type: object
            properties:
                detail:
                    type: string
...
        DiscountTier:
            type: object
            required: [level, quantity_kg, unit_price]
            properties:
                level:
                    type: integer
                    minimum: 1
                    description: Discount tier number
                quantity_kg:
                    type: number
                    format: float
                    description: "Quantity threshold [kg]"
                unit_price:
                    type: number
                    format: float
                    description: "Unit price [PLN/kg]"
        DailyPrice:
            type: object
            required: [day, base_price, availability_kg]
            properties:
                day:
                    type: integer
                    minimum: 1
                    description: Planning day
                base_price:
                    type: number
                    format: float
                    description: "Base price [PLN/kg]"
                availability_kg:
                    type: number
                    format: float
                    description: "Availability [kg]"
                discount_tiers:
                    type: array
                    items:
                        $ref: "#/components/schemas/DiscountTier"
                    description: "Discount tiers applicable on this day"
        DeliveryParams:
            type: object
            required: [building_id, lead_time_days, fixed_cost_pln]
            properties:
                building_id:
                    type: string
                    description: Building ID
                lead_time_days:
                    type: integer
                    minimum: 0
                    description: "Lead time [days]"
                fixed_cost_pln:
                    type: number
                    format: float
                    description: "Fixed delivery cost [PLN]"
                correction_cost_per_kg:
                    type: number
                    format: float
                    default: 0.0
                    description: "Correction cost [PLN/kg]"
                max_correction_kg:
                    type: number
                    format: float
                    default: 1000000.0
                    description: "Maximum correction [kg]"
        DistributorCreateRequest:
            type: object
            required: [username, contact_email, contact_phone, daily_prices, delivery_params]
            properties:
                username:
                    type: string
                contact_email:
                    type: string
                    format: email
                contact_phone:
                    type: string
                daily_prices:
                    type: array
                    items:
                        $ref: "#/components/schemas/DailyPrice"
                delivery_params:
                    type: array
                    items:
                        $ref: "#/components/schemas/DeliveryParams"
        DistributorUpdateRequest:
            type: object
            properties:
                username:
                    type: string
                    nullable: true
                contact_email:
                    type: string
                    nullable: true
                contact_phone:
                    type: string
                    nullable: true
                daily_prices:
                    type: array
                    nullable: true
                    items:
                        $ref: "#/components/schemas/DailyPrice"
                delivery_params:
                    type: array
                    nullable: true
                    items:
                        $ref: "#/components/schemas/DeliveryParams"
        DistributorResponse:
            allOf:
                - $ref: "#/components/schemas/DistributorCreateRequest"
                - type: object
                  properties:
                      id:
                          type: string
                      active:
                          type: boolean
                      created_at:
                          type: string
                          format: date-time
                      updated_at:
                          type: string
                          format: date-time
...
        InventoryStatus:
            type: object
            properties:
                building_id:
                    type: string
                    format: UUID
                building_name:
                    type: string
                current_inventory_kg:
                    type: number
                    format: float
                max_capacity_kg:
                    type: number
                    format: float
                fill_percent:
                    type: number
                    format: float
                    description: "Warehouse fill level [%]"
...